Task 2
Logos are a representation of what a brand or organization stands for. They speak volumes about what companies do and subtly express their values. Professionals often follow the Keep It Simple, Stupid (KISS) approach while designing logos, focusing on minimalism and the business’ core values. This fundamental principle is suggestive of the power of simplicity in graphic design.