Task 4
Infographics employ pictures, data visualizations, typography, and statistical graphics to tell a straightforward story or present information. Working on an infographic will give you experience with important tools like Canva and Venngage and allow you to exercise your creativity as you think of ways to represent information in a way that is visually attractive, impactful, and effective.